refactor(planner): move AIConfiguratorPerfEstimator to planner monitoring
Relocates the AIC wrapper from the profiler to the planner so the planner
can run AIC-based interpolation at bootstrap in rapid mode. The profiler's
estimate_perf.py becomes a re-export shim for existing callers.

Class identity is preserved — profiler and planner imports point at the
same underlying class, so isinstance checks still work.
